// src/ts/enums/ErrorMessages.ts
export enum ErrorMessages {
    SECTIONS_MISSING = "No section elements have been defined!",
    ONLY_ONE_PARENT_ELEMENT = "Exactly one element must have the parentClass",
    PARENT_MISSING = "The element with parentClass does not exist",
    ELEMENT_MISSING = "Elements with elementClass are not defined",
    TOP_WRAPPER_MISSING = "Top wrapper element is missing!",
    ELEMENT_WRAPPER_MISSING = "Element wrapper element is missing!",
    MAIN_WRAPPER_MISSING = "Main wrapper element is missing!",
}